chore: rename operator class
