+++ system/event-script-engine/src/main/java/com/accenture/services/plugins/collection/IsEmptyOperator.java
@@ -1,4 +1,4 @@
-package com.accenture.services.plugins.logical;
+package com.accenture.services.plugins.collection;
 
 import com.accenture.models.PluginFunction;
 import com.accenture.models.SimplePlugin;
@@ -18,15 +18,13 @@ public class IsEmptyOperator implements PluginFunction {
     @Override
     public Object calculate(Object... input) {
         if (input == null || input.length != 1) {
-            throw new IllegalArgumentException(
-                    "One input is required to check if value is empty");
+            throw new IllegalArgumentException("One input is required to check if value is empty");
         }
 
         Object value = input[0];
 
         if (value == null) {
-            throw new IllegalArgumentException(
-                    "Input cannot be null to check if value is empty");
+            throw new IllegalArgumentException("Input cannot be null to check if value is empty");
         }
 
         return switch (value) {
